moved switchable UI tests to another file
